refactor: move `usePosition` to `components/Gallery`
