+++ src/components/Gallery/composables/usePositions.ts
